Move all main actions and types into the main state domain
