+++ src/actions/batchedActions.js
@@ -22,7 +22,8 @@ import {
   globalLoadError,
   globalLoadFinish,
   globalLoadStart,
-} from 'stores/global/actions';
+} from 'stores/main/actions';
+import { getInfo } from 'stores/main/actions';
 import { modalHide } from 'stores/modal/actions';
 import {
   clusterNodePoolsLoad,
@@ -35,7 +36,7 @@ import {
   organizationsLoad,
 } from 'stores/organization/actions';
 import { loadReleases } from 'stores/releases/actions';
-import { getInfo, refreshUserInfo } from 'stores/user/actions';
+import { refreshUserInfo } from 'stores/user/actions';
 
 export const batchedLayout = () => async (dispatch) => {
   dispatch(globalLoadStart());
