+++ src/components/Cluster/ClusterDetail/ClusterDetailView.js
@@ -31,10 +31,10 @@ import {
   isClusterCreating,
   isClusterUpdating,
 } from 'stores/cluster/utils';
+import { getUserIsAdmin } from 'stores/main/selectors';
 import * as nodePoolActions from 'stores/nodepool/actions';
 import { NODEPOOL_MULTIPLE_LOAD_REQUEST } from 'stores/nodepool/constants';
 import { selectNodePools } from 'stores/nodepool/selectors';
-import { getUserIsAdmin } from 'stores/user/selectors';
 import SlideTransition from 'styles/transitions/SlideTransition';
 import Button from 'UI/Button';
 import ClusterIDLabel from 'UI/ClusterIDLabel';
