+++ src/reducers/index.js
@@ -6,6 +6,7 @@ import clusters from 'stores/cluster/reducer';
 import clusterLabels from 'stores/clusterlabels/reducer';
 import cpAuth from 'stores/cpauth/reducer';
 import errorReducer from 'stores/error/reducer';
+import makeMainReducer from 'stores/main/reducer';
 import metadata from 'stores/metadata/reducer';
 import modal from 'stores/modal/reducer';
 import nodePools from 'stores/nodepool/reducer';
@@ -16,7 +17,6 @@ import users from 'stores/user/reducer';
 import errorsByEntity from './entityErrorReducer';
 import loadingFlagsByEntity from './entityLoadingReducer';
 import loadingFlags from './loadingReducer';
-import makeMainReducer from './mainReducer';
 
 const entities = combineReducers({
   cpAuth: FeatureFlags.FEATURE_CP_ACCESS && cpAuth,
