+++ src/reducers/types.ts
@@ -1,9 +1,9 @@
-import { IInstallationInfo } from 'model/services/giantSwarm/types';
 import { IAppCatalogsState } from 'stores/appcatalog/types';
 import { IClusterState } from 'stores/cluster/types';
 import { IClusterLabelsState } from 'stores/clusterlabels/types';
 import { ICPAuthState } from 'stores/cpauth/types';
 import { IErrorState } from 'stores/error/types';
+import { IMainState } from 'stores/main/types';
 import { IMetadataState } from 'stores/metadata/types';
 import { IModalState } from 'stores/modal/types';
 import { INodePoolState } from 'stores/nodepool/types';
@@ -16,15 +16,7 @@ export interface IState {
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   [key: string]: any;
 
-  main: {
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    [key: string]: any;
-    loggedInUser: ILoggedInUser;
-    info: IInstallationInfo;
-    selectedOrganization: string | null;
-    firstLoadComplete: boolean;
-    selectedClusterID?: string;
-  };
+  main: IMainState;
   errors: IErrorState;
   metadata: IMetadataState;
   modal: IModalState;
