+++ src/stores/main/constants.ts
@@ -1,5 +1,5 @@
+export const CLUSTER_SELECT = 'CLUSTER_SELECT';
+
 export const GLOBAL_LOAD_REQUEST = 'GLOBAL_LOAD_REQUEST';
 export const GLOBAL_LOAD_SUCCESS = 'GLOBAL_LOAD_SUCCESS';
 export const GLOBAL_LOAD_ERROR = 'GLOBAL_LOAD_ERROR';
-
-export const CLUSTER_SELECT = 'CLUSTER_SELECT';
