+++ src/stores/main/reducer.ts
@@ -1,0 +1,119 @@
+import produce from 'immer';
+import { Providers } from 'shared/constants';
+import { GLOBAL_LOAD_ERROR, GLOBAL_LOAD_SUCCESS } from 'stores/main/constants';
+import { CLUSTER_SELECT } from 'stores/main/constants';
+import { IMainState, MainActions } from 'stores/main/types';
+import {
+  ORGANIZATION_LOAD_SUCCESS,
+  ORGANIZATION_SELECT,
+} from 'stores/organization/constants';
+import { OrganizationActions } from 'stores/organization/types';
+import {
+  INFO_LOAD_SUCCESS,
+  LOGIN_ERROR,
+  LOGIN_SUCCESS,
+  LOGOUT_ERROR,
+  LOGOUT_SUCCESS,
+  REFRESH_USER_INFO_ERROR,
+  REFRESH_USER_INFO_SUCCESS,
+} from 'stores/user/constants';
+import { UserActions } from 'stores/user/types';
+import {
+  fetchSelectedOrganizationFromStorage,
+  fetchUserFromStorage,
+  removeUserFromStorage,
+  setUserToStorage,
+} from 'utils/localStorageUtils';
+
+const initialState = (): IMainState => ({
+  selectedOrganization: fetchSelectedOrganizationFromStorage(),
+  selectedClusterID: undefined,
+  firstLoadComplete: false,
+  loggedInUser: fetchUserFromStorage(),
+  info: {
+    general: {
+      installation_name: '',
+      availability_zones: null,
+      provider: Providers.AWS,
+    },
+    stats: {
+      cluster_creation_duration: null,
+    },
+    workers: {
+      count_per_cluster: {
+        max: null,
+        default: 0,
+      },
+    },
+  },
+});
+
+const makeMainReducer = () => {
+  return produce(
+    (
+      draft: IMainState,
+      action: MainActions | UserActions | OrganizationActions
+    ) => {
+      switch (action.type) {
+        case REFRESH_USER_INFO_SUCCESS: {
+          const newUser = Object.assign({}, draft.loggedInUser, {
+            email: action.email,
+          });
+          setUserToStorage(newUser);
+
+          draft.loggedInUser = newUser;
+
+          break;
+        }
+
+        case INFO_LOAD_SUCCESS:
+          draft.info = action.info;
+
+          break;
+
+        case LOGIN_SUCCESS: {
+          setUserToStorage(action.userData);
+
+          draft.loggedInUser = action.userData;
+
+          break;
+        }
+
+        case REFRESH_USER_INFO_ERROR:
+        case LOGIN_ERROR:
+        case LOGOUT_SUCCESS:
+        case LOGOUT_ERROR:
+          removeUserFromStorage();
+
+          draft.loggedInUser = null;
+          draft.firstLoadComplete = false;
+
+          break;
+
+        case GLOBAL_LOAD_ERROR:
+        case GLOBAL_LOAD_SUCCESS:
+          draft.firstLoadComplete = true;
+
+          break;
+
+        case ORGANIZATION_SELECT:
+          draft.selectedOrganization = action.orgId;
+
+          break;
+
+        case ORGANIZATION_LOAD_SUCCESS:
+          draft.selectedOrganization = action.selectedOrganization;
+
+          break;
+
+        case CLUSTER_SELECT:
+          draft.selectedClusterID = action.clusterID;
+
+          break;
+      }
+    },
+    initialState()
+  );
+};
+
+export default makeMainReducer;
