+++ testUtils/preloginState.js
@@ -1,3 +1,4 @@
+import { Providers } from 'shared/constants';
 import { AppRoutes } from 'shared/constants/routes';
 
 // This is what the state looks like when someone brand new arrives at the site
@@ -15,15 +16,23 @@ export default {
   },
   main: {
     selectedOrganization: null,
+    selectedClusterID: undefined,
     firstLoadComplete: false,
     loggedInUser: null,
     info: {
       general: {
-        availability_zones: {
+        installation_name: '',
+        availability_zones: null,
+        provider: Providers.AWS,
+      },
+      stats: {
+        cluster_creation_duration: null,
+      },
+      workers: {
+        count_per_cluster: {
+          max: null,
           default: 0,
-          max: 0,
         },
-        provider: '',
       },
     },
   },
@@ -51,11 +60,6 @@ export default {
       isFetching: false,
       items: {},
       v5Clusters: [],
-    },
-    invitations: {
-      lastUpdated: 0,
-      isFetching: false,
-      items: {},
     },
     organizations: {
       lastUpdated: 0,
